backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.26.0
opencv-python-headless>=4.9.0.80
pytesseract>=0.3.10
pyspellchecker>=0.7.2
pypdfium2>=4.30.0
